LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.5_Regular_Expression_Literals/S7.8.5_A3.1_T9-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unexpected identifier '\u006D'. Expected ';' after variable declaration.
S7.8.5_A3.1_T9

FAIL successfullyParsed is not set

TEST COMPLETE
